refactor(classes):put middleware functions into classes
