+++ app/routes/user.js
@@ -1,10 +1,12 @@
-//event-prime\app\routes\eventcreator.js
+//event-prime\app\routes\user.js
 const express = require("express");
 const passport = require("passport");
 const session = require("express-session");
 
 const User = require("../controllers/user.js");
-const {signUp, signIn, createEvent} = require("../middleware/user.js");
+const Events = require("../controllers/events.js");
+const UserValidation = require("../middleware/uservalidation.js");
+const EventsValidation = require("../middleware/eventsvalidation.js");
 
 const app = express.Router();
 
@@ -12,9 +14,9 @@ app.use(session({resave:true, saveUninitialized:true, secret:"event-prime"}))
 app.use(passport.initialize());
 app.use(passport.session());
 
-app.put("/signup", signUp, User.addUser);
-app.post("/signin", signIn, User.logIn);
-app.put("/createevent", createEvent, User.addNewEvent);
+app.put("/signup", UserValidation.newUser, User.signUp);
+app.post("/signin", UserValidation.currentUser, User.signIn);
+app.put("/createevent", EventsValidation.newEvent, Events.addNewEvent);
 
 app.put("/welcome", (request, response)=>{
